+++ titiler/xarray/main.py
@@ -5,7 +5,9 @@ import os
 import shutil
 
 import rioxarray
+import zarr
 from fastapi import FastAPI
+from starlette import status
 from starlette.middleware.cors import CORSMiddleware
 
 import titiler.xarray.reader as reader
@@ -50,6 +52,10 @@ app.include_router(tms.router, tags=["Tiling Schemes"])
 algorithms = AlgorithmFactory()
 app.include_router(algorithms.router, tags=["Algorithms"])
 
+error_codes = {
+    zarr.errors.GroupNotFoundError: status.HTTP_422_UNPROCESSABLE_ENTITY,
+}
+add_exception_handlers(app, error_codes)
 add_exception_handlers(app, DEFAULT_STATUS_CODES)
 
 # Set all CORS enabled origins